style: 🎨 Change file structure
