+++ DrugClasses/index.js
@@ -1,0 +1,5 @@
+export { Drug } from "./Drug.js";
+export { HerbalTea } from "./HerbalTea.js";
+export { MagicPill } from "./MagicPill.js";
+export { Fervex } from "./Fervex.js";
+export { Dafalgan } from "./Dafalgan.js";
