+++ PharmacyClass/pharmacy.js
@@ -1,0 +1,68 @@
+export class Pharmacy {
+  constructor(drugs = []) {
+    this.drugs = drugs;
+  }
+  legacyUpdateBenefitValue() {
+    for (var i = 0; i < this.drugs.length; i++) {
+      if (
+        this.drugs[i].name != "Herbal Tea" &&
+        this.drugs[i].name != "Fervex"
+      ) {
+        if (this.drugs[i].benefit > 0) {
+          if (this.drugs[i].name != "Magic Pill") {
+            this.drugs[i].benefit = this.drugs[i].benefit - 1;
+          }
+        }
+      } else {
+        if (this.drugs[i].benefit < 50) {
+          this.drugs[i].benefit = this.drugs[i].benefit + 1;
+          if (this.drugs[i].name == "Fervex") {
+            if (this.drugs[i].expiresIn < 11) {
+              if (this.drugs[i].benefit < 50) {
+                this.drugs[i].benefit = this.drugs[i].benefit + 1;
+              }
+            }
+            if (this.drugs[i].expiresIn < 6) {
+              if (this.drugs[i].benefit < 50) {
+                this.drugs[i].benefit = this.drugs[i].benefit + 1;
+              }
+            }
+          }
+        }
+      }
+      if (this.drugs[i].name != "Magic Pill") {
+        this.drugs[i].expiresIn = this.drugs[i].expiresIn - 1;
+      }
+      if (this.drugs[i].expiresIn < 0) {
+        if (this.drugs[i].name != "Herbal Tea") {
+          if (this.drugs[i].name != "Fervex") {
+            if (this.drugs[i].benefit > 0) {
+              if (this.drugs[i].name != "Magic Pill") {
+                this.drugs[i].benefit = this.drugs[i].benefit - 1;
+              }
+            }
+          } else {
+            this.drugs[i].benefit =
+              this.drugs[i].benefit - this.drugs[i].benefit;
+          }
+        } else {
+          if (this.drugs[i].benefit < 50) {
+            this.drugs[i].benefit = this.drugs[i].benefit + 1;
+          }
+        }
+      }
+    }
+
+    return this.drugs;
+  }
+  updateBenefitValue() {
+    for (var i = 0; i < this.drugs.length; i++) {
+      if (this.drugs[i].name !== "Magic Pill") {
+        this.drugs[i].updateExpiresIn();
+        if (this.drugs[i].benefit < 50 && this.drugs[i].benefit > 0)
+          this.drugs[i].updateBenefit();
+      }
+    }
+    return this.drugs;
+  }
+}
